+++ Controls/_calendar/MonthList.ts
@@ -15,10 +15,9 @@ import ITEM_TYPES from './MonthList/ItemTypes';
 import {IDisplayedRanges, IDisplayedRangesOptions} from 'Controls/interface';
 import {IDateConstructor, IDateConstructorOptions} from 'Controls/interface';
 import {IDayTemplate, IDayTemplateOptions} from 'Controls/interface';
-import {IntersectionObserverSyntheticEntry} from 'Controls/scroll';
+import {IntersectionObserverSyntheticEntry, scrollToElement} from 'Controls/scroll';
 import {Base as dateUtils} from 'Controls/dateUtils';
 import {getDimensions} from 'Controls/sizeUtils';
-import {scrollToElement} from 'Controls/scrollUtils';
 import template = require('wml!Controls/_calendar/MonthList/MonthList');
 import monthTemplate = require('wml!Controls/_calendar/MonthList/MonthTemplate');
 import yearTemplate = require('wml!Controls/_calendar/MonthList/YearTemplate');
