+++ Controls/_scroll/ContainerBase.ts
@@ -4,8 +4,9 @@ import {Bus} from 'Env/Event';
 import {SyntheticEvent} from 'Vdom/Vdom';
 import {RegisterClass, Registrar} from 'Controls/event';
 import {Control, IControlOptions, TemplateFunction} from 'UI/Base';
-import {ResizeObserver as ResizeObserverUtil, scrollToElement} from 'Controls/scrollUtils';
+import ResizeObserverUtil from 'Controls/Utils/ResizeObserverUtil';
 import {canScrollByState, getContentSizeByState, getScrollPositionTypeByState, SCROLL_DIRECTION} from './Utils/Scroll';
+import {scrollToElement} from './Utils/scrollToElement';
 import {scrollTo} from './Utils/Scroll';
 import {IScrollState} from './Utils/ScrollState';
 import {SCROLL_MODE} from './Container/Type';
