+++ Controls/_scroll/StickyHeader/Controller.ts
@@ -3,7 +3,7 @@ import {debounce} from 'Types/function';
 import {IFixedEventData, isHidden, POSITION, SHADOW_VISIBILITY, TRegisterEventData, TYPE_FIXED_HEADERS} from './Utils';
 import StickyHeader from 'Controls/_scroll/StickyHeader';
 import fastUpdate from './FastUpdate';
-import {ResizeObserver as ResizeObserverUtil} from 'Controls/scrollUtils';
+import ResizeObserverUtil from 'Controls/Utils/ResizeObserverUtil';
 
 // @ts-ignore
