+++ Controls/_scroll/Utils/getScrollbarWidth.ts
